pyproject.toml
[tool.poetry]
name = "MeowerBot"
version = "2.4.0"
description = "A meower bot lib for py"
authors = ["showierdata9978 <68120127+showierdata9978@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
packages = [
    {include = "MeowerBot"}
]
repository = "https://github.com/MeowerBots/MeowerBot.py"


[tool.poetry.dependencies]
python = "^3.10"
websockets = "*"
aiohttp = "^3.8.4"
cleo = "^2.0.1"
requests = "^2.28.2"
pydotenv = "^0.0.7"

[tool.poetry.dev-dependencies]
black = "^22.6.0"

[tool.poetry.group.dev.dependencies]
pydotenv = "^0.0.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
